Bare bones welcome email
Created a bare bones welcome/registration email that currently sends
every time you log in. Will eventually change this to only send when a
user registers. To get this to work, you will need to set up your
environment variables MAIL_USERNAME and MAIL_PASSWORD; it is really
simple and directions can be found in the readme. I’ll also work on
some copy and an email template.