Click graph point, alert datapoint and label
As a user, I want to click a point in order to bring up congressional
records sorted by date desc. Records will be in a well to the right of
the graph. Will pre-populate this well with the most recent
congressional records with an occurence of this keyword. Tab separate
the two keywords in the well; will be able to toggle the tabs.